requirements.txt
bcrypt==5.0.0
camelsnake==0.0.2
PyPika==0.48.9
pytest==8.3.5
python-dotenv==1.0.1